Try adapting the pointers and errors section (#3)
* First example with the bitcoin wallet

Adapting the first few simple parts of the "Pointers and Errors" chapter
of "Learn Go With Tests" lead to some interesting things:

1. Tuple structs are how Rust lets you know something that is really
just an existing type
2. It doesn't seem like deriving `AddAssign` is possible, which is a
little disappointing, but, like with many things in Rust, there is
probably a good reason?
3. The biggest one is that, instead of getting the misleading assertion
failure like Go provides, Rust immediately complains at compile time
about moving the value if the deposit method doesn't use a mutable
reference to the wallet. That speaks volumes about how explicit Rust is
about which values are coming from where.

* Implement fmt::Display for Bitcoin

Implement a custom formatter for the first time really, then break the
test intentionally to see it working.

* Add withdraw method, refactor tests some

Implement `SubAssign` for the first time, then refactor some of the
tests, setting up the question of an overdraft next.

* Add Result return for withdrawal to handle insufficient funds

Since Rust insists on using Result to handle errors, these assertions
end up being really nice and clean and make a lot of sense. There is
still more to go to define an InsufficientFunds error of some kind, but
this is pretty nice.